requirements.txt
pip==22.3
click==8.1.3
colorama==0.4.4
Flask==2.0.2
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
Werkzeug==2.1.2
pillow==9.2.
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==1.11.0+cpu
pytorch-lightning==1.5.8
torchvision==0.12.0+cpu